requirements.txt
pydub
numpy
pandas
matplotlib
seaborn
scipy
Pillow
librosa
